Split ValueSnapshotter and IsolatableFactory impl so services can be at different scope
keeping the common implementation
and move ValueSnapshotter from user-home scope to build-session scope

and move DependencyManagementValueSnapshotterSerializerRegistry down to build-session scope where it will be able to use the dependency-management services required for supporting more types
