+++ subprojects/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/dsl/dependencies/GradlePluginVariantsSupportTest.groovy
@@ -24,7 +24,6 @@ import org.gradle.api.internal.attributes.EmptySchema
 import org.gradle.internal.component.model.AttributeMatchingExplanationBuilder
 import org.gradle.internal.component.model.AttributeSelectionSchema
 import org.gradle.internal.component.model.ComponentAttributeMatcher
-import org.gradle.internal.isolation.IsolatableFactory
 import org.gradle.util.AttributeTestUtil
 import org.gradle.util.SnapshotTestUtil
 import org.gradle.util.TestUtil
@@ -35,7 +34,7 @@ class GradlePluginVariantsSupportTest extends Specification {
     def attributes = AttributeTestUtil.attributesFactory()
     def objects = TestUtil.objectFactory()
     def matcher = new ComponentAttributeMatcher()
-    def schema = new DefaultAttributesSchema(matcher, TestUtil.instantiatorFactory(), SnapshotTestUtil.valueSnapshotter() as IsolatableFactory)
+    def schema = new DefaultAttributesSchema(matcher, TestUtil.instantiatorFactory(), SnapshotTestUtil.isolatableFactory())
     def ep = Stub(AttributeMatchingExplanationBuilder)
 
     def setup() {
