+++ subprojects/execution/src/integTest/groovy/org/gradle/internal/execution/IncrementalExecutionIntegrationTest.groovy
@@ -123,7 +123,7 @@ class IncrementalExecutionIntegrationTest extends Specification implements Valid
     }
     def outputSnapshotter = new DefaultOutputSnapshotter(snapshotter)
     def fingerprinterRegistry = new DefaultFileCollectionFingerprinterRegistry([FingerprinterRegistration.registration(DirectorySensitivity.DEFAULT, LineEndingSensitivity.DEFAULT, fingerprinter)])
-    def valueSnapshotter = new DefaultValueSnapshotter([], classloaderHierarchyHasher, null)
+    def valueSnapshotter = new DefaultValueSnapshotter([], classloaderHierarchyHasher)
     def inputFingerprinter = new DefaultInputFingerprinter(snapshotter, fingerprinterRegistry, valueSnapshotter)
     def buildCacheController = Mock(BuildCacheController)
     def buildOperationExecutor = new TestBuildOperationExecutor()
