+++ subprojects/model-core/src/test/groovy/org/gradle/api/internal/provider/ValueSourceBasedSpec.groovy
@@ -24,15 +24,14 @@ import org.gradle.api.provider.ValueSourceParameters
 import org.gradle.api.provider.ValueSourceSpec
 import org.gradle.internal.event.DefaultListenerManager
 import org.gradle.internal.service.scopes.Scopes
-import org.gradle.internal.snapshot.impl.DefaultValueSnapshotter
+import org.gradle.internal.snapshot.impl.DefaultIsolatableFactory
 import org.gradle.util.TestUtil
 import spock.lang.Specification
 
 abstract class ValueSourceBasedSpec extends Specification {
 
     def listenerManager = new DefaultListenerManager(Scopes.Build)
-    def isolatableFactory = new DefaultValueSnapshotter(
-        [],
+    def isolatableFactory = new DefaultIsolatableFactory(
         null,
         TestUtil.managedFactoryRegistry()
     )
